WIP: asyncInit
